[dnscrypt] Fix plain text TCP request.
We should only initialize a new dnscrypt_buffer when using the dnscrypt
listener. Otherwise, we just point dnscrypt_buffer to buffer.
The dnscrypt_buffer allocation is handled in services/listen_dnsport.c
`listen_create` when a new TCP/UDP listener is created and it is a
dnscrypt listener.
For TCP, we also create workers. By the time the workers are created, we don't
know if this is a dnscrypt worker or not, so instead, we differ
initialization of dnscrypt_buffer in the handler, if needed, we will allocate a
new buffer.